+++ tests/integration/functions/runtime_context/empty_undefined_parameters_test.ts
@@ -6,7 +6,7 @@ import { SlackFunctionTester } from "../../../../src/functions/tester/mod.ts";
 import { assertEqualsTypedValues } from "../../../../src/test_utils.ts";
 
 /**
- * Custom Function handler tests, exercising empty/undefined/lack of inputs/outputs as well as error or incomplete function return values
+ * Custom Function handler tests exercising empty/undefined/ inputs/outputs
  */
 Deno.test("Custom Function with no inputs or outputs", () => {
   const TestFn = DefineFunction({
@@ -61,54 +61,4 @@ Deno.test("Custom Function with empty inputs and outputs", () => {
   const { createContext } = SlackFunctionTester(TestFn);
   const result = handler(createContext({ inputs: {} }));
   assertEqualsTypedValues(result.outputs, {});
-});
-
-Deno.test("Custom Function that returns completed=false", () => {
-  const TestFn = DefineFunction({
-    callback_id: "test",
-    title: "test fn",
-    source_file: "test.ts",
-    output_parameters: {
-      properties: {
-        example: {
-          type: "boolean",
-        },
-      },
-      required: ["example"],
-    },
-  });
-  const handler: EnrichedSlackFunctionHandler<typeof TestFn.definition> =
-    () => {
-      return {
-        completed: false,
-      };
-    };
-  const { createContext } = SlackFunctionTester(TestFn);
-  const result = handler(createContext({ inputs: {} }));
-  assertEqualsTypedValues(result.completed, false);
-});
-
-Deno.test("Custom Function that returns error", () => {
-  const TestFn = DefineFunction({
-    callback_id: "test",
-    title: "test fn",
-    source_file: "test.ts",
-    output_parameters: {
-      properties: {
-        example: {
-          type: "string",
-        },
-      },
-      required: ["example"],
-    },
-  });
-  const handler: EnrichedSlackFunctionHandler<typeof TestFn.definition> =
-    () => {
-      return {
-        error: "error",
-      };
-    };
-  const { createContext } = SlackFunctionTester(TestFn);
-  const result = handler(createContext({ inputs: {} }));
-  assertEqualsTypedValues(result.error, "error");
 });
